resources_demo/requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
archtool==0.1.9
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fonttools==4.50.0
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
pyasn1==0.5.1
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-keycloak-async==2.17.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
